updated vendors management script
The VERSION file is gone as it was specific to Symfony, not to your application.

Installing vendors is as simple as running:

  ./bin/vendors install

This command installs all dependencies at a version known to work.

If you want to update all dependencies to their latest version,
you can run

  ./bin/vendors update

It updates your vendors and update the deps.lock file as well.
You should not have to run this command.