Refactoring the README to be very friendly to people after they've downloaded Symfony
I also renamed the file to .md, because GitHub doesn't render .rst very well, and I'm not aware of any place where we actually compile this README file using sphinx. I think this is just a pragmatic thing to do - if you look at the current README.rst on GitHub, some of the "code" blocks aren't processing the line breaks as desired.