I've removed PHP short tags, of the format <?= $foo ?> and replaced them with standard PHP echo tags: <?php echo $foo ?>
I've just spent half an hour trying to figure out why I was getting a blank login page, until I found out that short tags were disabled.  For an open-source project that supports PHP 5.3 and is (presumably) looking for a wide user-base, this feature cannot be relied upon, unfortunately.

I agree that it makes the code less neat, but I think that is the lesser of two evils if this is going to be a serious contender in the CI arena.