Cross-platform safe copy in Plugin.CopyBuild
Rather than using `xargs` which behaves differently on BSD and Linux, it's safer to use the straight `cp`. We lose the output of which files were copied - but I'd be more than happy to add this as a separate command to list which files were copied and which were ignored.