SSH
I'm currently running Windows machine and cygwin. This causes a hang because it requires user input. This causes the request just to go on forever until the connection is either closed or maximum execution time is reached.

So how I have thought to solve it is to add the help option.