feature #728 Added a new AppBundle to comply with Symfony Best Practices (javiereguiluz)
This PR was squashed before being merged into the 3.0-dev branch (closes #728).

Discussion
----------

Added a new AppBundle to comply with Symfony Best Practices

| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | yes
| Fixed tickets | #671
| License       | MIT
| Doc PR        | -

Symfony Best Practices recommend to start developing your applications with a single AppBundle bundle. This PR allows to include that AppBundle in the Symfony Standard Edition, to make it easier for developers.

I used the same skeleton as for the `generate:bundle` command, but I thought that it would be better to simplify it and remove the `$name` route parameter and use just a static template without parameters.

Commits
-------

45da0d0 Added a new AppBundle to comply with Symfony Best Practices