Make sure tests are run with full error reporting
This is also done in all symfony components. It makes sure that tests report all errors and warnings including strict ones.
Often this is not the case otherwise. The other advantage is that it makes tests more equal in PHP 5 and PHP 7. PHP  7 removed strict errors and changed them to warnings or notices. So in PHP 7 those errors might cause failures but they do not cause failures in PHP 5 when strict is ignored.