Properly lock down app_dev.php
Currently, web/app_dev.php in the symphony-standard package limits requests to those coming from CLI and unproxied localhost.
This prevents an attacker to execute this file in a production environment.
IPv4 localhost requests can come from 127.0.0.0/8, so the current check is actually a little bit too restrictive, but safe.
IPv6 localhost requests come from ::1/128.
The address fe80::1, however, is a link-local IPv6 address like any other. Anyone in the local network can take it and communicate with it. Therefore, it does not belong here.